bibliteca-django-main/requirements.txt
asgiref==3.8.1
certifi==2024.8.30
charset-normalizer==3.4.0
Django==4.2
django-model-utils==5.0.0
django-notifications-hq==1.8.3
django_notifications==0.1.dev0
idna==3.10
jsonfield==3.1.0
pillow==11.0.0
pytz==2024.2
requests==2.32.3
setuptools==75.5.0
sqlparse==0.5.2
swapper==1.4.0
urllib3==2.2.3
